[FIX] Resolve an issue where todo were not visible when moving from a constant to a scheduled todo
